add check models is ordered before deleting

+++ src/AppAdminBundle/Controller/CRUDController.php
@@ -70,6 +70,15 @@ class CRUDController extends BaseController
         );
     }
 
+    public function batchActionDeleteIsRelevant(array $selectedIds)
+    {
+        if (!$this->getDoctrine()->getRepository('AppBundle:ProductModelSpecificSize')
+            ->isProductModelsIsOrdered($selectedIds)){
+            return true;
+        }
+        return $this->admin->trans( 'flash_delete_not_ordered_error1', [ ], 'AppAdminBundle' );
+    }
+
     /**
      * Batch action activate
      *
@@ -203,7 +212,13 @@ class CRUDController extends BaseController
 
             return $this->redirectTo($object);
         }
-
+         if ($this->get('product')->checkProductsIsOrdered($object)){
+             $this->addFlash(
+                 'warning',
+                 $this->admin->trans( 'flash_delete_not_ordered_error1', [ ], 'AppAdminBundle' )
+             );
+             return $this->redirectTo($object);
+         }
         return parent::deleteAction($id);
     }
 }
